Refactoring improvement of the Person class
Improvement and creation of new wrapper allowing to manage mono valued
attributes nullable or not and mono valued attributes with a limited
size
Custom exception creation making errors more understandable.
Modifications of the test classes to take into account the new
implementation of the Person class, and modification of the names of the
test methods.

Resolves : #9