Refactoring of the Correspondence and City classes.
Use of wrappers for the different attributes of these classes to keep a
global consistency in the project.
Taking into account the cardinalities constraints of the UML diagram
between Correspondence and City.
Creation of a new wrapper Interval allowing to better manage the
handling of the dates of a Correspondence.

Resolves : #11
Resolves : #12